tests/requirements.txt
selene==2.0.0rc9
pytest
allure-pytest
python-dotenv==0.20.0